PDP update (#7)

+++ auth/kongUtils.py
@@ -1,8 +1,9 @@
 import logging
 import requests
 from requests import ConnectionError
+
 import conf
-from flaskAlchemyInit import HTTPRequestError
+from database.flaskAlchemyInit import HTTPRequestError
 
 LOGGER = logging.getLogger('device-manager.' + __name__)
 LOGGER.addHandler(logging.StreamHandler())
@@ -10,6 +11,14 @@ LOGGER.setLevel(logging.INFO)
 
 
 def configureKong(user):
+    # Disable Kong not advised. Only use for debug purposes
+    if conf.kongURL == 'DISABLED':
+        return {
+                'key': 'nokey',
+                'secret': 'nosecret',
+                'kongid': 'noid'
+                }
+
     try:
         exists = False
         response = requests.post('%s/consumers' % conf.kongURL,
@@ -44,6 +53,8 @@ def configureKong(user):
 
 # Invalidate old kong shared secret
 def revokeKongSecret(username, tokenId):
+    if conf.kongURL == 'DISABLED':
+        return
     try:
         requests.delete("%s/consumers/%s/jwt/%s"
                         % (conf.kongURL, username, tokenId))
@@ -53,6 +64,8 @@ def revokeKongSecret(username, tokenId):
 
 
 def removeFromKong(user):
+    if conf.kongURL == 'DISABLED':
+        return
     try:
         requests.delete("%s/consumers/%s" % (conf.kongURL, user))
     except ConnectionError:
